+++ ui/filters.go
@@ -1,0 +1,19 @@
+package ui
+
+/*
+Resets all filters of all types
+*/
+func ClearAllFilters() {
+
+}
+
+/*
+Resets filters of a specific type
+*/
+func ClearSpecificFilters() {
+
+}
+
+func AddFilter() {
+
+}